moved blits file type functions to lib
